requirements.txt
jieba==0.39
six==1.11.0
matplotlib==2.1.2
numpy==1.14.0
scikit-learn==0.19.1
requests==2.18.4
pyltp==0.2.1
pyahocorasick==1.1.8
flask==0.12.2